requirements.txt
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
Jinja2==2.10
